Added Game Over and New Level Generation
The game now ends when the player is touched by the enemy. If the
player gets the coin it displays that you won and then generates a new
level. I also added a file called ThingsToDo to keep track of what all
I need to do. I also did a little bit of code clean up. No
documentation yet.