Added Bullets and Rotation
The player can now fire bullets with the key ‘f’ and rotate using the
keys ‘q’ and ‘e’. The Bullet objects currently don’t collide with
anything or of anything. They also have a few bugs and the fps needs to
be updated across the entire game. I removed a bug on the level up
screen that caused an infinite loop and never changed.