Added Lantern
The Lantern is now in the game. There are still a few bugs in the game
but overall the lantern is working. It is not documented yet. The
Bullets you can currently see, the size of the lantern can’t be
changed, and the fuel level doesn’t decrease with time.